requirements-pqc.txt
liboqs-python>=0.10.0
